tools/akakce_worker/requirements.txt
beautifulsoup4==4.13.4
curl-cffi>=0.7,<1
Pillow==11.2.1
playwright==1.53.0
requests==2.32.3
